modbus-cliente/requirements.txt
psycopg2-binary
pymodbus
